Issue#23: include relevant XML service specification in generated API JARs
Including the relevant XML service specification file in a generated API
JAR provides a simple, definitive way to see the exact XML service
specification the generated stubs correspond to.

The XML service specification files are added in the directory 'xml', e.g.

    xml/ServiceDefMAL.xml

These files are not very big, and XML compresses very well, so the
increase in size is minimal.